requirements.txt
# PermitIQ - Python Dependencies
# ETL pipeline requirements

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.0.0

# Development dependencies (optional)
black>=23.0.0
pylint>=3.0.0
pytest>=7.4.0
